RoomService 수정에 따른 Test 코드 수정 및 MemberService 삭제에 따른 Test 코드 삭제
